Fix aks cluster provisioning command
When originally written, the command to provision an aks cluster
had to specify a specific version of Kubernetes. The requirement
of having Kubernetes version 1.13.x or higher is now met by the
default version of Kubernetes in all regions.

Additionally, all clusters now have rbac enabled by default so this
option is also removed.

https://github.com/dapr/docs/issues/832