Add lightweight analytics pixel to dapr website (#3025)
This analytics pixel provides high-level web traffic insights for project maintainers so they can better understand how docs are being read and used. This pixel does not use cookies (or JavaScript at all) and is GDPR-compliant. It is a clear 1x1 pixel that won't affect page rendering or load times.

Signed-off-by: Arjun Devarajan <arjun@scarf.sh>

Signed-off-by: Arjun Devarajan <arjun@scarf.sh>