Fix documentation for query parameters.
Moves moves query parameter from the API documentation (where it was
lost and bound to be forgotten and left out-of-sync to component
development) to each individual secret store documentation (which was
the right place to begin with and where there's a greater chance
of it being kept up-to-date with each component development).

Also fixes the examples that were malformed and that misled customers
to believe that Vault had support for a `namespace`  query parameter
as brought up in dapr/components-contrib#2806.

Fixes #3353

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>